Add support for sessionToken
I use this library on ec2 using instance roles. AWS creates temporary
credentials in this case which require a sessionToken. I modified this
library to allow an optional sessionToken.

I also changed the default Vertx.vertx() creation to not use a static
initializer. This was causing multiple vertx instances to be created when
used in an existing vertx project.